--- packages/workers/iframely/.dev.vars.example
@@ -1,1 +1,0 @@
-IFRAMELY_API_KEY=""
