--- packages/workers/iframely/README.md
@@ -1,1 +1,0 @@
-# IFramely worker
