--- packages/workers/iframely/src/types.ts
@@ -1,4 +1,0 @@
-export interface Env {
-  IFRAMELY_API_KEY: string;
-  LENSTER_IFRAMELY: KVNamespace;
-}
