--- packages/workers/iframely/wrangler.toml
@@ -1,15 +1,0 @@
-name = "iframely"
-main = "src/index.ts"
-compatibility_date = "2023-01-25"
-keep_vars = true
-
-routes = [
-	{ pattern = "iframely.lenster.xyz", custom_domain = true }
-]
-
-kv_namespaces = [
-  { binding = "LENSTER_IFRAMELY", id = "9c7a95492bbc43a89dc22cee5977cf1f", preview_id = "9c7a95492bbc43a89dc22cee5977cf1f" }
-]
-
-[env.production.vars]
-IFRAMELY_API_KEY = ""
