+++ packages/workers/oembed/README.md
@@ -1,0 +1,1 @@
+# Oembed worker
