+++ packages/workers/oembed/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@workers/iframely",
+  "name": "@workers/oembed",
   "version": "0.0.0",
   "private": true,
   "license": "GPL-3.0",
